lab/open-telemetry/pyproject.toml
[tool.poetry]
name = "otlp-tests"
version = "0.1.0"
description = ""
authors = ["Johnny Tardin <johnnytardin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
python-decouple = "^3.5"
gevent = "^21.12.0"
requests = "^2.28.1"
Faker = "^15.0.0"
install = "^1.3.5"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
